--- examples/tl_danet/config/movie_reg.yaml
@@ -1,15 +1,0 @@
-dataset: 'movielens_reg'
-task: 'regression'
-resume_dir: ''
-logname: 'layer32'
-
-fit:
-  max_epochs: 2
-  patience: 2000
-  lr: 0.22
-
-model:
-  layer: 2
-  base_outdim: 96
-  k: 8
-  drop_rate: 0.25
