--- examples/tl_danet/lib/utils.py
@@ -1,196 +1,0 @@
-from torch.utils.data import Dataset
-import torch
-import numpy as np
-from sklearn.utils import check_array
-
-class FastTensorDataLoader:
-    """
-    A DataLoader-like object for a set of tensors that can be much faster than
-    TensorDataset + DataLoader because dataloader grabs individual indices of
-    the dataset and calls cat (slow).
-    Source: https://discuss.pytorch.org/t/dataloader-much-slower-than-manual-batching/27014/6
-    """
-    def __init__(self, *tensors, batch_size=32, shuffle=False):
-        """
-        Initialize a FastTensorDataLoader.
-        :param *tensors: tensors to store. Must have the same length @ dim 0.
-        :param batch_size: batch size to load.
-        :param shuffle: if True, shuffle the data *in-place* whenever an
-            iterator is created out of this object.
-        :returns: A FastTensorDataLoader.
-        """
-        assert all(t.shape[0] == tensors[0].shape[0] for t in tensors)
-        self.tensors = tensors
-
-        self.dataset_len = self.tensors[0].shape[0]
-        self.batch_size = batch_size
-        self.shuffle = shuffle
-
-        # Calculate # batches
-        n_batches, remainder = divmod(self.dataset_len, self.batch_size)
-        if remainder > 0:
-            n_batches += 1
-        self.n_batches = n_batches
-    def __iter__(self):
-        if self.shuffle:
-            r = torch.randperm(self.dataset_len)
-            self.tensors = [t[r] for t in self.tensors]
-        self.i = 0
-        return self
-
-    def __next__(self):
-        if self.i >= self.dataset_len:
-            raise StopIteration
-        batch = tuple(t[self.i:self.i+self.batch_size] for t in self.tensors)
-        self.i += self.batch_size
-        return batch
-
-    def __len__(self):
-        return self.n_batches
-
-class PredictDataset(Dataset):
-    """
-    Format for numpy array
-
-    Parameters
-    ----------
-    X : 2D array
-        The input matrix
-    """
-
-    def __init__(self, x):
-        self.x = x
-
-    def __len__(self):
-        return len(self.x)
-
-    def __getitem__(self, index):
-        x = self.x[index]
-        return x
-
-def create_dataloaders(X_train, y_train, eval_set, batch_size):
-    """
-    Create dataloaders with or without subsampling depending on weights and balanced.
-
-    Parameters
-    ----------
-    X_train : np.ndarray
-        Training data
-    y_train : np.array
-        Mapped Training targets
-    eval_set : list of tuple
-        List of eval tuple set (X, y)
-    batch_size : int
-        how many samples per batch to load
-    Returns
-    -------
-    train_dataloader, valid_dataloader : torch.DataLoader, torch.DataLoader
-        Training and validation dataloaders
-    """
-    X_train = torch.from_numpy(X_train).float()
-    y_train = torch.from_numpy(y_train)
-    train_dataloader = FastTensorDataLoader(X_train, y_train, batch_size=batch_size, shuffle=True)
-
-    valid_dataloaders = []
-    for X, y in eval_set:
-        X = torch.from_numpy(X).float()
-        y = torch.from_numpy(y)
-        valid_dataloaders.append(FastTensorDataLoader(X, y, batch_size=batch_size, shuffle=False))
-
-    return train_dataloader, valid_dataloaders
-
-def validate_eval_set(eval_set, eval_name, X_train, y_train):
-    """Check if the shapes of eval_set are compatible with (X_train, y_train).
-
-    Parameters
-    ----------
-    eval_set : list of tuple
-        List of eval tuple set (X, y).
-        The last one is used for early stopping
-    eval_name : list of str
-        List of eval set names.
-    X_train : np.ndarray
-        Train owned products
-    y_train : np.array
-        Train targeted products
-
-    Returns
-    -------
-    eval_names : list of str
-        Validated list of eval_names.
-    eval_set : list of tuple
-        Validated list of eval_set.
-
-    """
-    eval_name = eval_name or [f"val_{i}" for i in range(len(eval_set))]
-
-    assert len(eval_set) == len(
-        eval_name
-    ), "eval_set and eval_name have not the same length"
-    if len(eval_set) > 0:
-        assert all(
-            len(elem) == 2 for elem in eval_set
-        ), "Each tuple of eval_set need to have two elements"
-    for name, (X, y) in zip(eval_name, eval_set):
-        check_array(X)
-        msg = (
-            f"Dimension mismatch between X_{name} "
-            + f"{X.shape} and X_train {X_train.shape}"
-        )
-        assert len(X.shape) == len(X_train.shape), msg
-
-        msg = (
-            f"Dimension mismatch between y_{name} "
-            + f"{y.shape} and y_train {y_train.shape}"
-        )
-        assert len(y.shape) == len(y_train.shape), msg
-
-        msg = (
-            f"Number of columns is different between X_{name} "
-            + f"({X.shape[1]}) and X_train ({X_train.shape[1]})"
-        )
-        assert X.shape[1] == X_train.shape[1], msg
-
-        if len(y_train.shape) == 2:
-            msg = (
-                f"Number of columns is different between y_{name} "
-                + f"({y.shape[1]}) and y_train ({y_train.shape[1]})"
-            )
-            assert y.shape[1] == y_train.shape[1], msg
-        msg = (
-            f"You need the same number of rows between X_{name} "
-            + f"({X.shape[0]}) and y_{name} ({y.shape[0]})"
-        )
-        assert X.shape[0] == y.shape[0], msg
-
-    return eval_name, eval_set
-
-def define_device(device_name):
-    """
-    Define the device to use during training and inference.
-    If auto it will detect automatically whether to use cuda or cpu
-
-    Parameters
-    ----------
-    device_name : str
-        Either "auto", "cpu" or "cuda"
-
-    Returns
-    -------
-    str
-        Either "cpu" or "cuda"
-    """
-    if device_name == "auto":
-        if torch.cuda.is_available():
-            return "cuda"
-        else:
-            return "cpu"
-    elif device_name == "cuda" and not torch.cuda.is_available():
-        return "cpu"
-    else:
-        return device_name
-
-def normalize_reg_label(label, mu, std):
-    norm_label = ((label - mu) / std).astype(np.float32)
-    norm_label = norm_label.reshape(-1, 1)
-    return norm_label
